turn creatId into long type

+++ localfullnode2/src/main/java/one/inve/localfullnode2/store/EventStoreDependency.java
@@ -34,7 +34,7 @@ public class EventStoreDependency implements EventStoreDependent, DependentItemC
 	}
 
 	@Override
-	public int getCreatorId() {
+	public long getCreatorId() {
 		return creatorId.get();
 	}
